[#652] Move test file to tests folder to try and make Travis happy
